big ole mesh shuffle
