+++ packages/mesh-extras/src/geometry/RopeGeometry.js
@@ -1,5 +1,4 @@
-import Mesh2d from './Mesh2d';
-
+import { MeshGeometry } from '@pixi/mesh';
 /**
  * The rope allows you to draw a texture across several points and them manipulate these points
  *
@@ -11,45 +10,47 @@ import Mesh2d from './Mesh2d';
  *  ```
  *
  * @class
- * @extends PIXI.Mesh
+ * @extends PIXI.MeshGeometry
  * @memberof PIXI
  *
  */
-export default class Rope extends Mesh2d
+export default class RopeGeometry extends MeshGeometry
 {
     /**
      * @param {PIXI.Texture} texture - The texture to use on the rope.
      * @param {PIXI.Point[]} points - An array of {@link PIXI.Point} objects to construct this rope.
      */
-    constructor(texture, points)
+    constructor(width = 200, points)
     {
-        super(texture, new Float32Array(points.length * 4),
+        super(new Float32Array(points.length * 4),
             new Float32Array(points.length * 4),
-            new Uint16Array(points.length * 2),
-            5);
+            new Uint16Array((points.length - 1) * 6));
 
         /*
          * @member {PIXI.Point[]} An array of points that determine the rope
          */
         this.points = points;
-        this.refresh();
+
+        this.width = width;
+
+        this.build();
     }
     /**
      * Refreshes Rope indices and uvs
      * @private
      */
-    _refresh()
+    build()
     {
         const points = this.points;
 
         if (!points) return;
 
-        const vertexBuffer = this.geometry.getAttribute('aVertexPosition');
-        const uvBuffer = this.geometry.getAttribute('aTextureCoord');
-        const indexBuffer = this.geometry.getIndex();
+        const vertexBuffer = this.getAttribute('aVertexPosition');
+        const uvBuffer = this.getAttribute('aTextureCoord');
+        const indexBuffer = this.getIndex();
 
         // if too little points, or texture hasn't got UVs set yet just move on.
-        if (points.length < 1 || !this.texture._uvs)
+        if (points.length < 1)
         {
             return;
         }
@@ -59,7 +60,7 @@ export default class Rope extends Mesh2d
         {
             vertexBuffer.data = new Float32Array(points.length * 4);
             uvBuffer.data = new Float32Array(points.length * 4);
-            indexBuffer.data = new Uint16Array(points.length * 2);
+            indexBuffer.data = new Uint16Array((points.length - 1) * 6);
         }
 
         const uvs = uvBuffer.data;
@@ -70,15 +71,15 @@ export default class Rope extends Mesh2d
         uvs[2] = 0;
         uvs[3] = 1;
 
-        indices[0] = 0;
-        indices[1] = 1;
+        // indices[0] = 0;
+        // indices[1] = 1;
 
-        const total = points.length;
+        const total = points.length; // - 1;
 
-        for (let i = 1; i < total; i++)
+        for (let i = 0; i < total; i++)
         {
             // time to do some smart drawing!
-            let index = i * 4;
+            const index = i * 4;
             const amount = i / (total - 1);
 
             uvs[index] = amount;
@@ -86,24 +87,34 @@ export default class Rope extends Mesh2d
 
             uvs[index + 2] = amount;
             uvs[index + 3] = 1;
+        }
 
-            index = i * 2;
-            indices[index] = index;
-            indices[index + 1] = index + 1;
+        let indexCount = 0;
+
+        for (let i = 0; i < total - 1; i++)
+        {
+            const index = i * 2;
+
+            indices[indexCount++] = index;
+            indices[indexCount++] = index + 1;
+            indices[indexCount++] = index + 2;
+
+            indices[indexCount++] = index + 2;
+            indices[indexCount++] = index + 1;
+            indices[indexCount++] = index + 3;
         }
 
         // ensure that the changes are uploaded
         uvBuffer.update();
         indexBuffer.update();
 
-        this.multiplyUvs();
-        this.refreshVertices();
+        this.updateVertices();
     }
 
     /**
      * refreshes vertices of Rope mesh
      */
-    refreshVertices()
+    updateVertices()
     {
         const points = this.points;
 
@@ -119,7 +130,7 @@ export default class Rope extends Mesh2d
 
         // this.count -= 0.2;
 
-        const vertices = this.vertices;
+        const vertices = this.buffers[0].data;
         const total = points.length;
 
         for (let i = 0; i < total; i++)
@@ -147,7 +158,7 @@ export default class Rope extends Mesh2d
             }
 
             const perpLength = Math.sqrt((perpX * perpX) + (perpY * perpY));
-            const num = this._texture.height / 2; // (20 + Math.abs(Math.sin((i + this.count) * 0.3) * 50) )* ratio;
+            const num = this.width / 2; // (20 + Math.abs(Math.sin((i + this.count) * 0.3) * 50) )* ratio;
 
             perpX /= perpLength;
             perpY /= perpLength;
@@ -163,20 +174,11 @@ export default class Rope extends Mesh2d
             lastPoint = point;
         }
 
-        this.geometry.buffers[0].update();
+        this.buffers[0].update();
     }
 
-    /**
-     * Updates the object transform for rendering
-     *
-     * @private
-     */
-    updateTransform()
+    update()
     {
-        if (this.autoUpdate)
-        {
-            this.refreshVertices();
-        }
-        this.containerUpdateTransform();
+        this.updateVertices();
     }
 }
