+++ packages/mesh-extras/src/index.js
@@ -1,4 +1,8 @@
-export { default as Mesh2d } from './Mesh2d';
-export { default as Plane } from './Plane';
-export { default as NineSlicePlane } from './NineSlicePlane';
-export { default as Rope } from './Rope';
+// export { default as Mesh2d } from './Mesh2d';
+// export { default as SimplePlane } from './SimplePlane';
+export { default as PlaneGeometry } from './geometry/PlaneGeometry';
+export { default as RopeGeometry } from './geometry/RopeGeometry';
+export { default as SimpleRope } from './SimpleRope';
+export { default as SimplePlane } from './SimplePlane';
+export { default as SimpleMesh } from './SimpleMesh';
+
