--- packages/mesh-extras/src/mesh.frag
@@ -1,9 +1,0 @@
-varying vec2 vTextureCoord;
-uniform vec4 uColor;
-
-uniform sampler2D uSampler;
-
-void main(void)
-{
-    gl_FragColor = texture2D(uSampler, vTextureCoord) * uColor;
-}
