+++ packages/mesh/src/index.js
@@ -2,5 +2,3 @@ export { default as Mesh } from './Mesh';
 export { default as MeshMaterial } from './MeshMaterial';
 export { default as MeshGeometry } from './MeshGeometry';
 
-export { default as PlaneGeometry } from './geometry/PlaneGeometry';
-export { default as RopeGeometry } from './geometry/RopeGeometry';
